agentic-repos-program/5- Chainlit_conversational_ AI_app_using_Gemini/conversation_agent/pyproject.toml
[project]
name = "conversation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
